Refactor: do not create callbacks in render() fn.
