+++ src/js/components/header/login.js
@@ -17,6 +17,14 @@ type Props = {
   disabled: boolean,
   nubbinPosition: string,
 };
+type MenuOption =
+  | {|
+      label: string,
+      href?: string,
+      disabled: boolean,
+      modal?: boolean,
+    |}
+  | {| type: string |};
 
 class Login extends React.Component<Props, { modalOpen: boolean }> {
   static defaultProps = {
@@ -46,8 +54,20 @@ class Login extends React.Component<Props, { modalOpen: boolean }> {
     this.setState({ modalOpen: isOpen });
   };
 
-  render(): React.Node {
-    const menuOpts = [
+  handleSelect = (opt: MenuOption) => {
+    if (opt.modal) {
+      this.toggleModal(true);
+      return;
+    }
+    if (opt.href) {
+      window.location.assign(
+        addUrlParams(opt.href, { next: window.location.pathname }),
+      );
+    }
+  };
+
+  static getMenuOpts(): Array<MenuOption> {
+    return [
       {
         label: 'Production or Developer Org',
         href:
@@ -71,6 +91,10 @@ class Login extends React.Component<Props, { modalOpen: boolean }> {
         disabled: !window.api_urls.salesforce_custom_login,
       },
     ];
+  }
+
+  render(): React.Node {
+    const menuOpts = Login.getMenuOpts();
     return (
       <>
         <Dropdown
@@ -88,17 +112,7 @@ class Login extends React.Component<Props, { modalOpen: boolean }> {
           iconName="down"
           iconPosition="right"
           options={menuOpts}
-          onSelect={opt => {
-            if (opt.modal) {
-              this.toggleModal(true);
-              return;
-            }
-            if (opt.href) {
-              window.location.assign(
-                addUrlParams(opt.href, { next: window.location.pathname }),
-              );
-            }
-          }}
+          onSelect={this.handleSelect}
         />
         <CustomDomainModal
           isOpen={this.state.modalOpen}
