+++ src/js/components/plans/stepsTable/kindDataCell.js
@@ -1,0 +1,35 @@
+// @flow
+
+import * as React from 'react';
+import DataTableCell from '@salesforce/design-system-react/components/data-table/cell';
+import Icon from '@salesforce/design-system-react/components/icon';
+
+import type { DataCellProps } from 'components/plans/stepsTable/index';
+
+const KindDataCell = (props: DataCellProps): React.Node => {
+  /* istanbul ignore if */
+  if (!props.item) {
+    return null;
+  }
+  const value = props.item.kind;
+  const iconName = props.item.kind_icon;
+  return (
+    <DataTableCell title={value} {...props}>
+      {iconName ? (
+        <Icon
+          className="slds-m-right_x-small"
+          category="utility"
+          name={iconName}
+          assistiveText={{
+            label: value,
+          }}
+          size="x-small"
+        />
+      ) : null}
+      <span>{value}</span>
+    </DataTableCell>
+  );
+};
+KindDataCell.displayName = DataTableCell.displayName;
+
+export default KindDataCell;
